Add new various NixOS configurations and utilities
This commit includes the addition of new NixOS configurations and utilities. It also creates and adjusts several pre-commit hooks and builds diverse modules to be used in different contexts. Moreover, this commit updates the user reference in keys.nix. Lastly, it initiates the creation of a vast array of ISO images with customized configurations.