Update README.md and modify repository references
This commit updates the instructions in the README.md file and modifies references to external repositories in the flake.nix file. The documentation is now more straightforward with a detailed setup guide. It also changes the old references to Repositories "RoboNyx" and "RoboNyx-lib" to "Robo-Cowboys/modules" and "Robo-Cowboys/lib" respectively on GitHub, making the repository sources more accurate.